webapp/backend/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
slowapi
bcrypt<=4.0.1
jwt